requirements.txt
pyyaml==3.12
requests==2.18.4
pytest==3.6.4
docker==3.4.1
docker-compose==1.22.0
kubernetes==6.0
#maestro-ng==0.5.4
